pyproject.toml
[tool.poetry]
name = "metastock"
version = "0.1.1"
description = ""
authors = ["wsl <mr.vjcspy@gmail.com>"]
readme = "README.md"


[tool.poetry.scripts]
pystock = "metastock.main:app"

[tool.poetry.dependencies]
python = "^3.11"
annotated-types = "^0.5.0"
arrow = "^1.2.3"
attrs = "^23.1.0"
certifi = "^2023.7.22"
charset-normalizer = "^3.2.0"
click = "^8.1.7"
colorama = "^0.4.6"
contourpy = "^1.1.0"
cycler = "^0.11.0"
fonttools = "^4.42.1"
idna = "^3.4"
jsonschema = "^4.19.0"
jsonschema-specifications = "^2023.7.1"
kiwisolver = "^1.4.5"
markdown-it-py = "^3.0.0"
marshmallow = "^3.20.1"
matplotlib = "^3.7.2"
mdurl = "^0.1.2"
mplfinance = "^0.12.10b0"
numpy = "^1.25.2"
packaging = "^23.1"
pandas = "^2.0.3"
pendulum = "^2.1.2"
pika = "^1.3.2"
pillow = "^10.0.0"
pydantic = "^2.3.0"
pydantic-core = "^2.6.3"
pygments = "^2.16.1"
pyparsing = "^3"
python-dateutil = "^2.8.2"
python-dotenv = "^1.0.0"
pytz = "^2023.3"
pytzdata = "^2020.1"
referencing = "^0.30.2"
requests = "^2.31.0"
rich = "^13.5.2"
rpds-py = "^0.9.2"
shellingham = "^1.5.3"
simplejson = "^3.19.1"
six = "^1.16.0"
termcolor = "^2.3.0"
typer = "^0.9.0"
typing-extensions = "^4.7.1"
tzdata = "^2023.3"
urllib3 = "^2.0.4"
pytest = "^7.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
